Rename Crud controller to habits to keep with convention
